+++ core/jvm/main/java/com/webank/ai/eggroll/core/grpc/client/GrpcClientTemplate.java
@@ -18,11 +18,11 @@ package com.webank.ai.eggroll.core.grpc.client;
 
 import com.google.protobuf.Message;
 import com.webank.ai.eggroll.core.concurrent.AwaitSettableFuture;
+import com.webank.ai.eggroll.core.error.handler.DefaultLoggingErrorHandler;
+import com.webank.ai.eggroll.core.error.handler.ErrorHandler;
+import com.webank.ai.eggroll.core.error.handler.InterruptAndRethrowRuntimeErrorHandler;
 import com.webank.ai.eggroll.core.factory.GrpcStreamComponentFactory;
-import com.webank.eggroll.core.error.handler.DefaultLoggingErrorHandler;
-import com.webank.eggroll.core.error.handler.ErrorHandler;
-import com.webank.eggroll.core.error.handler.InterruptAndRethrowRuntimeErrorHandler;
-import com.webank.eggroll.core.util.ErrorUtils;
+import com.webank.ai.eggroll.core.util.ErrorUtils;
 import io.grpc.stub.AbstractStub;
 import io.grpc.stub.StreamObserver;
 import java.util.concurrent.CountDownLatch;
