+++ core/jvm/main/scala/com/webank/ai/eggroll/core/error/handler/ErrorHandler.scala
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.webank.ai.eggroll.core.error.handler
+
+import com.webank.ai.eggroll.core.constant.StringConstants
+import com.webank.ai.eggroll.core.util.Logging
+
+trait ErrorHandler {
+  def handleError(t: Throwable): Unit = handleError(StringConstants.EMPTY, t)
+
+  def handleError(msg: String, t: Throwable): Unit
+}
+
+case class DefaultLoggingErrorHandler() extends ErrorHandler with Logging {
+  override def handleError(msg: String, t: Throwable): Unit = {
+    logError(msg, t)
+  }
+}
+
+case class InterruptAndRethrowRuntimeErrorHandler() extends ErrorHandler with Logging {
+  override def handleError(msg: String, t: Throwable): Unit = {
+    logError(msg, t)
+    Thread.currentThread().interrupt()
+    throw new RuntimeException(t)
+  }
+}
