--- core/jvm/main/scala/com/webank/eggroll/core/session/GrpcServerConf.scala
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.webank.eggroll.core.session
-
-import com.webank.eggroll.core.model.Endpoint
-import io.grpc.ServerServiceDefinition
-
-abstract class GrpcServerConf(confFilePath: String) {
-  var endpoint: Endpoint
-  var serverServiceDefinition: java.util.List[ServerServiceDefinition]
-  var isSecureClient: Boolean
-  var isSecureServer: Boolean
-}
