--- core/jvm/test/java/com/webank/eggroll/core/di/TestDi.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.webank.eggroll.core.di;
-
-import com.webank.ai.eggroll.core.di.Singletons;
-import java.util.ArrayList;
-import org.junit.Test;
-
-public class TestDi {
-
-  @Test
-  public void getList() {
-    ArrayList list = Singletons.getNoCheck(ArrayList.class);
-    assert (list != null);
-
-    ArrayList list2 = Singletons.getNoCheck(ArrayList.class);
-    assert (list == list2);
-  }
-}
