--- core/jvm/test/scala/com/webank/eggroll/core/factory/TestGrpcStreamComponentFactory.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.webank.eggroll.core.factory;
-
-import com.webank.ai.eggroll.core.di.Singletons;
-import com.webank.ai.eggroll.core.factory.GrpcStreamComponentFactory;
-import org.junit.Test;
-
-public class TestGrpcStreamComponentFactory {
-
-  private GrpcStreamComponentFactory grpcStreamComponentFactory = Singletons.getNoCheck(
-      GrpcStreamComponentFactory.class);
-
-  @Test
-  public void testCreate() throws Exception {
-
-/*    serverFactory.equals(null);
-    assert(serverFactory != null);*/
-  }
-}
